found a major vulnerability with manual collapse. removed the feature from the smart contract. renamed many smart contract methods to the alias-names in this codebase
