+++ src/types.ts
@@ -141,6 +141,14 @@ export type PeriodInfo = {
   claimIds: readonly bigint[];
 };
 
+export type ManualCollapseInfo = {
+  isManuallyCollapsed: false;
+} | {
+  isManuallyCollapsed: true;
+  initiatedTimestamp: bigint;
+  cancelDeadlineTimestamp: bigint;
+}
+
 /** enum type representing the different states the TandaPay community can be in */
 export enum TandaPayState {
   /**
