+++ test/helpers/tandapay_test_suite.ts
@@ -1,5 +1,6 @@
 import { WriteableTandaPayManager } from "tandapay_manager/tandapay_manager";
 import {
+  AssignmentStatus,
   memberInfoJsonReplacer,
   MemberStatus,
   subgroupInfoJsonReplacer,
@@ -30,7 +31,8 @@ import {
   DEFAULT_COVERAGE_REQUIREMENT,
   DEFAULT_CLAIMANT_INDEX,
 } from "../test_config";
-import { TandaPayLog, toTandaPayLogs } from "tandapay_manager/read/types";
+import { toTandaPayLogs } from "tandapay_manager/read/types";
+import { TandaPayLog } from "tandapay_manager/read/types";
 
 export type TestClientWithPublicActions = TestClient & PublicClient;
 export type DoActionForEachManagerParams = {
@@ -386,7 +388,8 @@ export class TandaPayTestSuite {
       memberInfo = await m.read.getMemberInfoFromAddress(address);
       // they should become reorged at this time
       expect(memberInfo.memberStatus).toBe(MemberStatus.Reorged);
-
+      
+      //console.log(`before self-approval:\n\tmember status: ${MemberStatus[memberInfo.memberStatus]}\n\tassignment status: ${AssignmentStatus[memberInfo.assignmentStatus]}`);
       // they approve their own subgroup assignment
       await m.write.member.approveSubgroupAssignment(true);
 
@@ -396,13 +399,18 @@ export class TandaPayTestSuite {
         fail(
           `no peer was populated in the subgroupPeers map for subgroup ID ${newSubgroupId}`,
         );
-
+     
+      //TODO: remove
+      //memberInfo = await m.read.getMemberInfoFromAddress(address);
+      //console.log(`before peer-approval:\n\tmember status: ${MemberStatus[memberInfo.memberStatus]}\n\tassignment status: ${AssignmentStatus[memberInfo.assignmentStatus]}`);
+      //TODO: end removal
       await peer.write.member.approveNewSubgroupMember(
         newSubgroupId,
         memberInfo.id,
         true,
       );
       memberInfo = await m.read.getMemberInfoFromAddress(address);
+      //console.log(`after peer-approval:\n\tmember status: ${MemberStatus[memberInfo.memberStatus]}\n\tassignment status: ${AssignmentStatus[memberInfo.assignmentStatus]}`);
       // now they should go back to being valid, since they've been assigned to a new subgroup successfully
       expect(memberInfo.memberStatus).toBe(MemberStatus.Valid);
     }
